+++ strata/src/trie/mpt/mod.rs
@@ -2,6 +2,6 @@ mod mutation;
 pub(crate) mod proof;
 mod query;
 
-pub use mutation::TrieMut;
+pub use mutation::{MAX_MPT_KEY_LEN, TrieMut};
 pub use proof::MptProof;
 pub use query::TrieRo;